extract wave source check route
