+++ src/api/routers/waves.py
@@ -25,7 +25,6 @@ from src.api.routers.wave_response_contracts import (
 from src.api.routers.wave_request_models import (
     DpmWaveSelectionRequest,
     DpmWaveSimulationRequest,
-    DpmWaveSourceCheckRequest,
     DpmWaveWorkflowCommandRequest,
 )
 from src.api.routers.wave_campaign_definition_http import (
@@ -47,15 +46,15 @@ from src.api.routers.wave_campaign_readiness_routes import (
     router as campaign_readiness_router,
 )
 from src.api.routers.wave_create_preview_routes import register_wave_create_preview_routes
+from src.api.routers.wave_source_check_routes import (
+    router as source_check_router,
+)
 from src.api.routers.wave_route_parameters import (
     CampaignDefinitionIdPath,
     CampaignDefinitionVersionPath,
     WaveCorrelationIdHeader,
     WaveIdPath,
     WaveItemIdPath,
-)
-from src.api.routers.wave_openapi_examples import (
-    SOURCE_CHECK_WAVE_EXAMPLE,
 )
 from src.api.routers.wave_read_http import (
     get_wave_detail_response,
@@ -66,7 +65,6 @@ from src.api.routers.wave_report_input_http import get_wave_report_input_respons
 from src.api.routers.wave_search_http import search_waves_response
 from src.api.routers.wave_selection_http import select_wave_item_alternative_response
 from src.api.routers.wave_simulation_http import simulate_wave_response
-from src.api.routers.wave_source_check_http import source_check_wave_response
 from src.api.routers.wave_supportability_http import get_wave_supportability_response
 from src.api.routers.wave_workflow_command_http import run_wave_workflow_command_response
 from src.api.routers.rebalance_runs import get_dpm_run_support_service
@@ -270,79 +268,7 @@ def get_wave_items(
     )
 
 
-@router.post(
-    "/{wave_id}/source-check",
-    response_model=DpmWaveResponse,
-    status_code=status.HTTP_200_OK,
-    summary="Source-check a durable rebalance wave",
-    description=(
-        "Evaluates RFC-0041 source readiness for each durable wave item using persisted mandate "
-        "digital twins, mandate health snapshots, and their source lineage. Items are classified "
-        "as `SOURCE_READY`, `SOURCE_DEGRADED`, `REVIEW_REQUIRED`, or `SOURCE_BLOCKED`; caller "
-        "portfolio ids or supplied refs alone never promote an item to ready. Repeating the call "
-        "after the wave is already `SOURCE_CHECKED` returns the persisted wave as an idempotent "
-        "replay without appending a duplicate event."
-    ),
-    responses={
-        200: {
-            "description": "Durable source-checked wave with item classifications.",
-            "content": {"application/json": {"example": SOURCE_CHECK_WAVE_EXAMPLE}},
-        },
-        404: {
-            "description": "Wave not found.",
-            "content": {
-                "application/json": {
-                    "example": {
-                        "detail": {
-                            "code": "DPM_WAVE_NOT_FOUND",
-                            "message": "Wave dwv_missing was not found.",
-                        }
-                    }
-                }
-            },
-        },
-        409: {
-            "description": "Wave version conflict during optimistic update.",
-            "content": {
-                "application/json": {
-                    "example": {
-                        "detail": {
-                            "code": "DPM_WAVE_VERSION_CONFLICT",
-                            "message": "DPM_WAVE_VERSION_CONFLICT",
-                        }
-                    }
-                }
-            },
-        },
-        422: {
-            "description": "Wave is not in a state that can be source-checked.",
-            "content": {
-                "application/json": {
-                    "example": {
-                        "detail": {
-                            "code": "DPM_WAVE_SOURCE_CHECK_INVALID_STATE",
-                            "message": "Wave dwv_001 cannot be source-checked from state DRAFT.",
-                        }
-                    }
-                }
-            },
-        },
-    },
-)
-def source_check_wave(
-    wave_id: WaveIdPath,
-    request: DpmWaveSourceCheckRequest,
-    x_correlation_id: WaveCorrelationIdHeader = None,
-    mandate_repository: DpmMandateRepository = Depends(get_mandate_repository),
-    wave_repository: DpmWaveRepository = Depends(get_wave_repository),
-) -> DpmWaveResponse:
-    return source_check_wave_response(
-        wave_id=wave_id,
-        request=request,
-        correlation_id=x_correlation_id or f"corr_wave_source_check_{wave_id}",
-        mandate_repository=mandate_repository,
-        wave_repository=wave_repository,
-    )
+router.include_router(source_check_router)
 
 
 @router.post(
